New Lagny solver
The Lagny solver was modified to more closely resemble Robin Leroy's
method in Kerbal Space Program.

It is still missing some pieces, such as:

* Kahan-like initialization (I use a Halley step from 0.7)

* Steve Canon's coefficients in the Lagny irrational step
  (I use the "exact" theoretical values.)

I'm sure there's more, but that's where I'm up to.
The main thing for now is that it does generate correct answers.

The speed is fine (if slower than Halley no-div).  Accuracy so far is no
different.

The bit correction section is a disaster, and I'm sure that I have no
idea what I'm doing.

(Also, some updates to the notes, nothing major though.)