Exact cbrt in Fortran!
The "final" cbrt in Fortran now gives the same "exact" result as
cbrt_ac.

The issue was that the r2_l "residual" would never use FMAs natively.  I
could only coax this to use FMAs with the explicit ieee_arithmetic
function.

This just never seems capable of doing the FMA; it always wants to
substitute r*r with r2_h.

    r2_h = r*r
    r2_l = r*r - r2_h

This was the only way to get it to work.

    r2_h = r*r
    r2_h = ieee_fma(r,r,-r2_h)

Perhaps there's some way, but it can't be dirty with, e.g. volatile
variables or something like that.  It has to be platform-independent and
optimization-favorable.  Currently, I see no alternative.